+++ backend/app/models/purchase.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import date
+from decimal import Decimal
 
 from sqlalchemy import Date, ForeignKey, Numeric, String, UniqueConstraint
 from sqlalchemy.dialects.postgresql import UUID
@@ -20,7 +21,7 @@ class Purchase(UUIDPKMixin, TimestampMixin, Base):
     )
     purchase_date: Mapped[date] = mapped_column(Date, nullable=False)
     invoice_no: Mapped[str | None] = mapped_column(String(100), nullable=True)
-    total_amount: Mapped[float] = mapped_column(Numeric(14, 2), nullable=False, default=0)
+    total_amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False, default=0)
 
     supplier = relationship("Supplier", backref="purchases")
     line_items = relationship(
